# application/api-spec/src/main/resources/openapi/nutrition-planner.yaml
openapi: 3.0.3
info:
  title: Nutrition Planner
  description: API for Nutrition Planner application
  version: 1.0.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /meal-plans:
    get:
      summary: List meal plans
      operationId: readMealPlans
      tags:
        - MealPlan
      responses:
        '200':
          description: Meal plans returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MealPlan'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new meal plan
      operationId: createMealPlan
      tags:
        - MealPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMealPlanRequest'
      responses:
        '201':
          description: Meal plan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MealPlan'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /meal-plans/{mealPlanId}:
    get:
      summary: Read meal plan by id
      operationId: readMealPlan
      tags:
        - MealPlan
      parameters:
        - $ref: '#/components/parameters/MealPlanId'
      responses:
        '200':
          description: Meal plan returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MealPlan'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Meal plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update meal plan
      operationId: updateMealPlan
      tags:
        - MealPlan
      parameters:
        - $ref: '#/components/parameters/MealPlanId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMealPlanRequest'
      responses:
        '200':
          description: Meal plan updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MealPlan'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Meal plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete meal plan
      operationId: deleteMealPlan
      tags:
        - MealPlan
      parameters:
        - $ref: '#/components/parameters/MealPlanId'
      responses:
        '204':
          description: Meal plan deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Meal plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /meal-plans/{mealPlanId}/activate:
    put:
      summary: Activate a meal plan
      operationId: activateMealPlan
      tags:
        - MealPlan
      parameters:
        - $ref: '#/components/parameters/MealPlanId'
      responses:
        '200':
          description: Meal plan activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MealPlan'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Meal plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /meal-plans/{mealPlanId}/deactivate:
    put:
      summary: Deactivate a meal plan
      operationId: deactivateMealPlan
      tags:
        - MealPlan
      parameters:
        - $ref: '#/components/parameters/MealPlanId'
      responses:
        '200':
          description: Meal plan deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MealPlan'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Meal plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /meal-plans/{mealPlanId}/days/{dayId}/entries:
    post:
      summary: Add entry to a plan day
      operationId: addPlanEntry
      tags:
        - MealPlan
      parameters:
        - $ref: '#/components/parameters/MealPlanId'
        - $ref: '#/components/parameters/DayId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPlanEntryRequest'
      responses:
        '201':
          description: Entry added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanEntry'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Meal plan or day not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /meal-plans/{mealPlanId}/days/{dayId}/entries/{entryId}:
    delete:
      summary: Remove entry from a plan day
      operationId: removePlanEntry
      tags:
        - MealPlan
      parameters:
        - $ref: '#/components/parameters/MealPlanId'
        - $ref: '#/components/parameters/DayId'
        - $ref: '#/components/parameters/EntryId'
      responses:
        '204':
          description: Entry removed
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    MealPlanId:
      name: mealPlanId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    DayId:
      name: dayId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    EntryId:
      name: entryId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
...
    MealTypeEnum:
      type: string
      enum:
        - BREAKFAST
        - LUNCH
        - DINNER
        - SNACK
    EntryTypeEnum:
      type: string
      enum:
        - MEAL
        - FOOD_PRODUCT
    CreateMealPlanRequest:
      type: object
      required:
        - name
        - startDate
        - numberOfDays
      properties:
        name:
          type: string
          minLength: 1
          pattern: "^.*\\S.*$"
        startDate:
          type: string
          format: date
        numberOfDays:
          type: integer
          minimum: 1
          maximum: 30
    UpdateMealPlanRequest:
      type: object
      required:
        - name
        - startDate
        - numberOfDays
      properties:
        name:
          type: string
          minLength: 1
          pattern: "^.*\\S.*$"
        startDate:
          type: string
          format: date
        numberOfDays:
          type: integer
          minimum: 1
          maximum: 30
    AddPlanEntryRequest:
      type: object
      required:
        - mealType
        - entryType
      properties:
        mealType:
          $ref: '#/components/schemas/MealTypeEnum'
        entryType:
          $ref: '#/components/schemas/EntryTypeEnum'
        mealId:
          type: integer
          format: int64
        portions:
          type: number
          format: double
          minimum: 0.0001
        foodProductId:
          type: integer
          format: int64
        grams:
          type: number
          format: double
          minimum: 0.0001
    PlanEntry:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        mealType:
          $ref: '#/components/schemas/MealTypeEnum'
        entryType:
          $ref: '#/components/schemas/EntryTypeEnum'
        mealId:
          type: integer
          format: int64
        mealName:
          type: string
        portions:
          type: number
          format: double
        foodProductId:
          type: integer
          format: int64
        foodProductName:
          type: string
        grams:
          type: number
          format: double
        calories:
          type: number
          format: double
        protein:
          type: number
          format: double
        fat:
          type: number
          format: double
        carbohydrates:
          type: number
          format: double
        price:
          type: number
          format: double
        sodiumMg:
          type: number
          format: double
        potassiumMg:
          type: number
          format: double
        magnesiumMg:
          type: number
          format: double
        ironMg:
          type: number
          format: double
        calciumMg:
          type: number
          format: double
        zincMg:
          type: number
          format: double
        vitaminAMcg:
          type: number
          format: double
        vitaminCMg:
          type: number
          format: double
        vitaminDMcg:
          type: number
          format: double
        vitaminEMg:
          type: number
          format: double
        vitaminKMcg:
          type: number
          format: double
        vitaminB1Mg:
          type: number
          format: double
        vitaminB2Mg:
          type: number
          format: double
        vitaminB6Mg:
          type: number
          format: double
        vitaminB9Mcg:
          type: number
          format: double
        vitaminB12Mcg:
          type: number
          format: double
    PlanDay:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        dayNumber:
          type: integer
        date:
          type: string
          format: date
        entries:
          type: array
          items:
            $ref: '#/components/schemas/PlanEntry'
        dailyCalories:
          type: number
          format: double
        dailyProtein:
          type: number
          format: double
        dailyFat:
          type: number
          format: double
        dailyCarbohydrates:
          type: number
          format: double
        dailyPrice:
          type: number
          format: double
        dailySodiumMg:
          type: number
          format: double
        dailyPotassiumMg:
          type: number
          format: double
        dailyMagnesiumMg:
          type: number
          format: double
        dailyIronMg:
          type: number
          format: double
        dailyCalciumMg:
          type: number
          format: double
        dailyZincMg:
          type: number
          format: double
        dailyVitaminAMcg:
          type: number
          format: double
        dailyVitaminCMg:
          type: number
          format: double
        dailyVitaminDMcg:
          type: number
          format: double
        dailyVitaminEMg:
          type: number
          format: double
        dailyVitaminKMcg:
          type: number
          format: double
        dailyVitaminB1Mg:
          type: number
          format: double
        dailyVitaminB2Mg:
          type: number
          format: double
        dailyVitaminB6Mg:
          type: number
          format: double
        dailyVitaminB9Mcg:
          type: number
          format: double
        dailyVitaminB12Mcg:
          type: number
          format: double
    MealPlan:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
        startDate:
          type: string
          format: date
        numberOfDays:
          type: integer
        isActive:
          type: boolean
        activatedAt:
          type: string
          format: date
        lastDeductedDayNumber:
          type: integer
        days:
          type: array
          items:
            $ref: '#/components/schemas/PlanDay'
        totalCalories:
          type: number
          format: double
        totalProtein:
          type: number
          format: double
        totalFat:
          type: number
          format: double
        totalCarbohydrates:
          type: number
          format: double
        totalPrice:
          type: number
          format: double
        totalSodiumMg:
          type: number
          format: double
        totalPotassiumMg:
          type: number
          format: double
        totalMagnesiumMg:
          type: number
          format: double
        totalIronMg:
          type: number
          format: double
        totalCalciumMg:
          type: number
          format: double
        totalZincMg:
          type: number
          format: double
        totalVitaminAMcg:
          type: number
          format: double
        totalVitaminCMg:
          type: number
          format: double
        totalVitaminDMcg:
          type: number
          format: double
        totalVitaminEMg:
          type: number
          format: double
        totalVitaminKMcg:
          type: number
          format: double
        totalVitaminB1Mg:
          type: number
          format: double
        totalVitaminB2Mg:
          type: number
          format: double
        totalVitaminB6Mg:
          type: number
          format: double
        totalVitaminB9Mcg:
          type: number
          format: double
        totalVitaminB12Mcg:
          type: number
          format: double
...